extra features and security

+++ CMS_TEMPLATE/admin/includes/view_all_users.php
@@ -70,12 +70,17 @@ if(isset($_GET['change_to_sub'])){
 
 
 //DELETING USERS
+
 if(isset($_GET['delete'])){
-    $the_user_id = $_GET['delete'];
+    if(isset($_SESSION['user_role'])){
+    if($_SESSION['user_role'] == 'admin') {    
+    
+    $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
     
     $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
     $delete_user_query = mysqli_query($connection, $query);
-
+    }
+    }
  header("Location: users.php");
 }
 ?>
